allow float

+++ schema.sql
@@ -15,7 +15,7 @@ CREATE TABLE IF NOT EXISTS metrics (
 CREATE TABLE IF NOT EXISTS data (
     metrics_id    INT UNSIGNED NOT NULL,
     datetime     DATETIME NOT NULL,
-    number       BIGINT NOT NULL,
+    number       REAL NOT NULL,
     updated_at   TIMESTAMP NOT NULL,
     PRIMARY KEY (metrics_id, datetime),
     KEY (datetime)
